# src/Umbraco.Cms.Api.Management/Controllers/DocumentType/Tree/ItemsDocumentTypeTreeController.cs
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Umbraco.Cms.Core.Services;
using Umbraco.Cms.Api.Management.ViewModels.Tree;

namespace Umbraco.Cms.Api.Management.Controllers.DocumentType.Tree;

public class ItemsDocumentTypeTreeController : DocumentTypeTreeControllerBase
{
    public ItemsDocumentTypeTreeController(IEntityService entityService, IContentTypeService contentTypeService)
        : base(entityService, contentTypeService)
    {
    }

    [HttpGet("item")]
    [MapToApiVersion("1.0")]
    [ProducesResponseType(typeof(IEnumerable<DocumentTypeTreeItemResponseModel>), StatusCodes.Status200OK)]
    public async Task<ActionResult<IEnumerable<DocumentTypeTreeItemResponseModel>>> Items([FromQuery(Name = "key")] Guid[] keys)
        => await GetItems(keys);
}