# src/Umbraco.Cms.Api.Management/OpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Umbraco Backoffice API",
    "description": "This shows all APIs available in this version of Umbraco - including all the legacy apis that are available for backward compatibility",
    "version": "1.0"
  },
  "paths": {
...
    "/umbraco/management/api/v1/audit-log/type/{logType}": {
      "get": {
        "tags": [
          "Audit Log"
        ],
        "operationId": "GetAuditLogTypeByLogType",
        "parameters": [
          {
            "name": "logType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AuditTypeModel"
            }
          },
          {
            "name": "sinceDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAuditLogResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/umbraco/management/api/v1/data-type/{key}": {
      "get": {
        "tags": [
          "Data Type"
        ],
        "operationId": "GetDataTypeByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DataTypeResponseModel"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Data Type"
        ],
        "operationId": "DeleteDataTypeByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Data Type"
        ],
        "operationId": "PutDataTypeByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateDataTypeRequestModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/umbraco/management/api/v1/tree/document-type/item": {
      "get": {
        "tags": [
          "Document Type"
        ],
        "operationId": "GetTreeDocumentTypeItem",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/DocumentTypeTreeItemResponseModel"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/umbraco/management/api/v1/document/{key}": {
      "get": {
        "tags": [
          "Document"
        ],
        "operationId": "GetDocumentByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DocumentResponseModel"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Document"
        ],
        "operationId": "DeleteDocumentByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Document"
        ],
        "operationId": "PutDocumentByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateDocumentRequestModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/umbraco/management/api/v1/tree/document/item": {
      "get": {
        "tags": [
          "Document"
        ],
        "operationId": "GetTreeDocumentItem",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "dataTypeKey",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "culture",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/DocumentTreeItemResponseModel"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/umbraco/management/api/v1/relation/type/{key}": {
      "get": {
        "tags": [
          "Relation"
        ],
        "operationId": "GetRelationTypeByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedRelationResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/umbraco/management/api/v1/template/{key}": {
      "get": {
        "tags": [
          "Template"
        ],
        "operationId": "GetTemplateByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TemplateResponseModel"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Template"
        ],
        "operationId": "DeleteTemplateByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Template"
        ],
        "operationId": "PutTemplateByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateTemplateRequestModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditLogBaseModel": {
        "type": "object",
        "properties": {
          "userKey": {
            "type": "string",
            "format": "uuid"
          },
          "entityKey": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "logType": {
            "$ref": "#/components/schemas/AuditTypeModel"
          },
          "entityType": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuditLogResponseModel": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/AuditLogBaseModel"
          }
        ],
        "additionalProperties": false
      },
...
      "AuditTypeModel": {
        "enum": [
          "New",
          "Save",
          "SaveVariant",
          "Open",
          "Delete",
          "Publish",
          "PublishVariant",
          "SendToPublish",
          "SendToPublishVariant",
          "Unpublish",
          "UnpublishVariant",
          "Move",
          "Copy",
          "AssignDomain",
          "PublicAccess",
          "Sort",
          "Notify",
          "System",
          "RollBack",
          "PackagerInstall",
          "PackagerUninstall",
          "Custom",
          "ContentVersionPreventCleanup",
          "ContentVersionEnableCleanup"
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ContentResponseModelBaseDocumentValueModelDocumentVariantResponseModel": {
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DocumentValueModel"
                }
              ]
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DocumentVariantResponseModel"
                }
              ]
            }
          },
          "key": {
            "type": "string",
            "format": "uuid"
          },
          "contentTypeKey": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "ContentTreeItemResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/EntityTreeItemResponseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "noAccess": {
            "type": "boolean"
          },
          "isTrashed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "ContentTreeItemResponseModel": "#/components/schemas/ContentTreeItemResponseModel",
            "DocumentTreeItemResponseModel": "#/components/schemas/DocumentTreeItemResponseModel"
          }
        }
      },
...
      "ContentUpdateRequestModelBaseDocumentValueModelDocumentVariantRequestModel": {
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DocumentValueModel"
                }
              ]
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DocumentVariantRequestModel"
                }
              ]
            }
          }
        },
        "additionalProperties": false
      },
      "ContentUrlInfoModel": {
        "type": "object",
        "properties": {
          "culture": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "DataTypeModelBaseModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "propertyEditorAlias": {
            "type": "string"
          },
          "propertyEditorUiAlias": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DataTypePropertyPresentationModel"
                }
              ]
            }
          }
        },
        "additionalProperties": false
      },
      "DataTypePropertyPresentationModel": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string"
          },
          "value": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataTypeResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/DataTypeModelBaseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "format": "uuid"
          },
          "parentKey": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "DataTypeResponseModel": "#/components/schemas/DataTypeResponseModel"
          }
        }
      },
...
      "DocumentResponseModel": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ContentResponseModelBaseDocumentValueModelDocumentVariantResponseModel"
          }
        ],
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ContentUrlInfoModel"
                }
              ]
            }
          },
          "templateKey": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DocumentTreeItemResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ContentTreeItemResponseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "isProtected": {
            "type": "boolean"
          },
          "isPublished": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "DocumentTreeItemResponseModel": "#/components/schemas/DocumentTreeItemResponseModel"
          }
        }
      },
...
      "DocumentTypeTreeItemResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/FolderTreeItemResponseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "isElement": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "DocumentTypeTreeItemResponseModel": "#/components/schemas/DocumentTypeTreeItemResponseModel"
          }
        }
      },
      "DocumentValueModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ValueModelBaseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "DocumentValueModel": "#/components/schemas/DocumentValueModel"
          }
        }
      },
      "DocumentVariantRequestModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/VariantModelBaseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "DocumentVariantRequestModel": "#/components/schemas/DocumentVariantRequestModel"
          }
        }
      },
      "DocumentVariantResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/VariantResponseModelBaseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ContentStateModel"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "DocumentVariantResponseModel": "#/components/schemas/DocumentVariantResponseModel"
          }
        }
      },
      "EntityTreeItemResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/TreeItemPresentationModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "format": "uuid"
          },
          "isContainer": {
            "type": "boolean"
          },
          "parentKey": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "EntityTreeItemResponseModel": "#/components/schemas/EntityTreeItemResponseModel",
            "ContentTreeItemResponseModel": "#/components/schemas/ContentTreeItemResponseModel",
            "DocumentBlueprintTreeItemResponseModel": "#/components/schemas/DocumentBlueprintTreeItemResponseModel",
            "DocumentTreeItemResponseModel": "#/components/schemas/DocumentTreeItemResponseModel",
            "DocumentTypeTreeItemResponseModel": "#/components/schemas/DocumentTypeTreeItemResponseModel",
            "FolderTreeItemResponseModel": "#/components/schemas/FolderTreeItemResponseModel"
          }
        }
      },
...
      "FolderTreeItemResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/EntityTreeItemResponseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "isFolder": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "FolderTreeItemResponseModel": "#/components/schemas/FolderTreeItemResponseModel",
            "DocumentTypeTreeItemResponseModel": "#/components/schemas/DocumentTypeTreeItemResponseModel"
          }
        }
      },
...
      "PagedAuditLogResponseModel": {
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/AuditLogResponseModel"
                }
              ]
            }
          }
        },
        "additionalProperties": false
      },
...
      "PagedRelationResponseModel": {
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RelationResponseModel"
                }
              ]
            }
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetailsModel": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RelationResponseModel": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "integer",
            "format": "int32"
          },
          "parentName": {
            "type": "string",
            "nullable": true
          },
          "childId": {
            "type": "integer",
            "format": "int32"
          },
          "childName": {
            "type": "string",
            "nullable": true
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TemplateModelBaseModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TemplateResponseModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/TemplateModelBaseModel"
          }
        ],
        "properties": {
          "$type": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "TemplateResponseModel": "#/components/schemas/TemplateResponseModel"
          }
        }
      },
...
      "UpdateDataTypeRequestModel": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/DataTypeModelBaseModel"
          }
        ],
        "additionalProperties": false
      },
...
      "UpdateDocumentRequestModel": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ContentUpdateRequestModelBaseDocumentValueModelDocumentVariantRequestModel"
          }
        ],
        "properties": {
          "templateKey": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTemplateRequestModel": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/TemplateModelBaseModel"
          }
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "OAuth": {
        "type": "oauth2",
        "description": "Umbraco Authentication",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/umbraco/management/api/v1.0/security/back-office/authorize",
            "tokenUrl": "/umbraco/management/api/v1.0/security/back-office/token",
            "scopes": { }
          }
        }
      }
    }
  },
  "security": [
    {
      "OAuth": [ ]
    }
  ]
}